+++ icevision/backbones/__init__.py
@@ -1,6 +1,10 @@
 from icevision.backbones.vgg import *
 from icevision.backbones.resnet import *
+from icevision.backbones.resnet_fpn import *
+from icevision.backbones.resnet_fpn_utils import *
+from icevision.backbones.resnest_fpn import *
 from icevision.backbones.mobilenet import *
 
+# TODO: Deprecate
 import icevision.backbones.resnet_fpn
 import icevision.backbones.resnest_fpn