+++ icevision/models/mmdet/common/bbox/two_stage/model.py
@@ -5,34 +5,29 @@ from mmcv import Config
 from mmdet.models import build_detector
 from mmcv.runner import load_checkpoint
 from icevision.models.mmdet.utils import *
+from icevision.models.mmdet.common.utils import *
 
 
 def model(
-    cfg: Union[str, Path, Config],
+    backbone: MMDetBackboneConfig,
     num_classes: int,
-    weights_path: Optional[Union[str, Path]] = None,
+    pretrained: bool = True,
+    checkpoints_path: Optional[Union[str, Path]] = "checkpoints",
+    force_download=False,
 ) -> nn.Module:
 
-    # if `cfg` argument is a path (str, Path) create an Config object from the file
-    # otherwise cfg should be already an Config object
-    if isinstance(cfg, (str, Path)):
-        cfg = Config.fromfile(str(cfg))
+    cfg, weights_path = create_model_config(
+        backbone=backbone,
+        pretrained=pretrained,
+        checkpoints_path=checkpoints_path,
+        force_download=force_download,
+    )
 
-    # Sparse-RCNN has a list of bbox_head whereas Faster-RCNN has only one
-    if isinstance(cfg.model.roi_head.bbox_head, list):
-        for bbox_head in cfg.model.roi_head.bbox_head:
-            bbox_head["num_classes"] = num_classes - 1
-    else:
-        cfg.model.roi_head.bbox_head.num_classes = num_classes - 1
-
-    if weights_path is not None:
-        cfg.model.pretrained = None
-
-    _model = build_detector(cfg.model, cfg.get("train_cfg"), cfg.get("test_cfg"))
-
-    if weights_path is not None:
-        load_checkpoint(_model, str(weights_path))
-
-    _model.param_groups = MethodType(param_groups, _model)
-
-    return _model
+    return build_model(
+        model_type="two_stage_detector_bbox",
+        backbone=backbone,
+        num_classes=num_classes,
+        pretrained=pretrained,
+        checkpoints_path=checkpoints_path,
+        force_download=force_download,
+    )