+++ icevision/models/ross/efficientdet/utils.py
@@ -1,0 +1,10 @@
+__all__ = [
+    "EfficientDetBackboneConfig",
+]
+
+from icevision.imports import *
+
+
+@dataclass
+class EfficientDetBackboneConfig:
+    model_name: str