+++ icevision/visualize/utils.py
@@ -1,7 +1,15 @@
-__all__ = ["draw_label", "bbox_polygon", "draw_mask"]
+__all__ = [
+    "draw_label",
+    "bbox_polygon",
+    "draw_mask",
+    "as_rgb_tuple",
+    "get_default_font",
+]
 
 from icevision.imports import *
 from matplotlib import patches
+from PIL import Image, ImageFont, ImageDraw
+import PIL
 
 
 def draw_label(ax, x, y, name, color, fontsize=18):
@@ -26,3 +34,39 @@ def draw_mask(ax, mask, color):
     color_mask = np.ones((*mask.shape, 3)) * color
     ax.imshow(np.dstack((color_mask, mask * 0.5)))
     ax.contour(mask, colors=[color_mask[0, 0, :]], alpha=0.4)
+
+
+def as_rgb_tuple(x: Union[np.ndarray, tuple, list, str]) -> tuple:
+    "Convert np RGB values -> tuple for PIL compatibility"
+    if isinstance(x, (np.ndarray, tuple, list)):
+        if not len(x) == 3:
+            raise ValueError(f"Expected 3 (RGB) numbers, got {len(x)}")
+        if isinstance(x, np.ndarray):
+            return tuple(x.astype(np.int))
+        elif isinstance(x, tuple):
+            return x
+        elif isinstance(x, list):
+            return tuple(x)
+    elif isinstance(x, str):
+        return PIL.ImageColor.getrgb(x)
+    else:
+        raise ValueError(f"Expected {{np.ndarray|list|tuple}}, got {type(x)}")
+
+
+def get_default_font() -> Path:
+    import requests
+
+    icedir = Path.home() / ".icevision"
+    icedir.mkdir(exist_ok=True)
+
+    font_dir = icedir / "fonts"
+    font_dir.mkdir(exist_ok=True)
+
+    font_file = font_dir / "SpaceGrotesk-Medium.ttf"
+    if not font_file.exists():
+        URL = "https://raw.githubusercontent.com/airctic/storage/master/SpaceGrotesk-Medium.ttf"
+        print(
+            f"Downloading default `.ttf` font file - SpaceGrotesk-Medium.ttf from {URL} to {font_file}"
+        )
+        font_file.write_bytes(requests.get(URL).content)
+    return str(font_file)