+++ tests/parsers/test_parser_all_components.py
@@ -19,12 +19,12 @@ def dummy_parser_all_components():
                     IsCrowdsRecordComponent(),
                 )
             )
-            super().__init__(record=record)
+            super().__init__(template_record=record)
 
         def __iter__(self) -> Any:
             raise NotImplementedError
 
-        def parse_fields(self, o, record: BaseRecord) -> None:
+        def parse_fields(self, o, record: BaseRecord, is_new) -> None:
             record.set_filepath(__file__)
             record.set_img_size(ImgSize(480, 420))
 
@@ -43,13 +43,13 @@ def test_parser_parse_fields(dummy_parser_all_components):
     parser = dummy_parser_all_components()
 
     record = parser.create_record()
-    parser.parse_fields(None, record)
+    parser.parse_fields(None, record, is_new=True)
 
     assert record.filepath == Path(__file__)
     assert record.height == 420
     assert record.width == 480
 
-    assert record.detection.labels == [1]
+    assert record.detection.label_ids == [1]
     assert record.detection.bboxes == [BBox.from_xyxy(1, 2, 3, 4)]
     assert record.detection.masks.erles == [{"size": [420, 480], "counts": b"PlT6"}]
     assert record.detection.areas == [4.2]